chore: 인터페이스 주석 추가 (스펙 설명)
