+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/domain/recipe/message/RecipeLikeMessagePublisher.java
@@ -1,6 +1,6 @@
 package com.mars.app.domain.recipe.message;
 
-import com.mars.app.config.rabbitmq.RecipeLikeRabbitConfig;
+import com.mars.app.config.rabbitmq.impl.RecipeLikeRabbitConfig;
 import com.mars.app.domain.recipe.dto.RecipeLikeMessageDto;
 import com.mars.app.domain.recipe.dto.RecipeLikeResponseDto;
 import lombok.RequiredArgsConstructor;
